Fixed search on transaction page
You cannot search the SQL database for python class property values. As
a result, trying to search over gear__name was giving a lookup error.
I replaced the this with a search over gear_data and geartype__name.
Behavior is not perfect, since this search does not match value+suffix
pairs, but still works basically.