Email host settings were set to send emails to localhost not smtp
When porting settings to heroku, it seems we ended up with the dev
version of the email settings. Therefore, emails have simply been
printed to console instead of getting sent to the ionos smtp server to
be forwarded to the actual recipient. Updated settings to pull from the
env vars